[PATCH] ppc64: IOMMU cleanups: rename pci_dma_direct.c
This patch renames pci_dma_direct.c to pci_direct_iommu.c to comply to the
naming convention of the other iommu files.

This is part of the iommu cleanup, but broken out as a separate patch since
for mainline, a BK rename is more appropriate.  Still, we need a patch to
apply for non-BK-based trees (-mm)

Signed-off-by: Olof Johansson <olof@austin.ibm.com>
Acked-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e0533b9jTZRI8Ft70AaUPkfzoVMQ