[PATCH] oprofile: fix ia64 callgraph bug with old gcc
With Keith Owens <kaos@sgi.com>

This patch from Keith Owens fixes a bug in the ia64 port of oprofile when
built without the kdb patch and with a pre-3.4 gcc.

If you build a standard kernel with gcc < 3.4 then
ia64_spinlock_contention_pre3_4 is defined.  But a standard kernel does not
have ia64_spinlock_contention_pre3_4_end, that label is only added by the
kdb patch.  To get the backtrace profiling with gcc < 3.4, the _end label
needs to be added as part of the kernprof patch, then I will remove it from
kdb.

Signed-off-by: Keith Owens <kaos@sgi.com>
Signed-off-by: Greg Banks <gnb@melbourne.sgi.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e05403fHc83heRG_9l2CqEbvAUPQ