[PATCH] Fix smp_processor_id() warning in numa_node_id()
The patch below fixes smp_processor_id() warnings that are triggered by
numa_node_id().

All uses of numa_node_id() in mm/mempolicy.c seem to use it as a 'hint'
only, not as a correctness number.  Once a node is established, it's used
in a preemption-safe way.  So the simple fix is to disable the checking for
numa_node_id().  But additional review would be more than welcome, because
this patch turns off the preemption-checking of numa_node_id() permanently.
Tested on amd64.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e056870B3elWzoKqHG6U1JJCPdOg