[PATCH] ppc64: fix iommu cleanup regression
In the recent IOMMU cleanup, the new LPAR code assumes that all PHBs must
have a dma window assigned to it.  On some machines we don't have a window
assinged unless there's an adapter in the slot.

In other words, a PHB without a ibm,dma-window property is not a bug and
must be tolerated.  This patch fixes that, and also removes a redundant
check for the dma-window being defined.

Signed-off-by: Olof Johansson <olof@austin.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e408aaTUUONGP3iB5O1YWwl_rwjA