[PATCH] ppc64: rename perf counter register #defines
This patch makes some cleanups to the #defines for various fields in the
MMCR0 performance monitor control register.  Specifically, the names of a
couple of bits are changed so that: a) they are a bit less cumbersomely
long and b) they match the names used in the hardware documentation.

Signed-off-by: David Gibson <dwg@au1.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e408b7bB6WbhwjHJFa48ZcvzROhg