[PATCH] random: remove pool resizing sysctl
Remove random pool resizing sysctl.  It's racy in hard to fix ways and
doesn't really warrant fixing.  It also only allows adjusting the input
pool size so it's either obsolete or incomplete.

The default 4kbits (512 bytes) of entropy pool should be more than big
enough for all purposes and too small to be a memory concern.

This patch removes the resizing code and marks the sysctl read-only.

Signed-off-by: Matt Mackall <mpm@selenic.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e40977kRMDKFAursNVBnBufBYZKA