[PATCH] reiserfs vs-8115 test adjustment
This patch below is needed to suppress the warning, with more than one
indirect block insertion at a time, it is now possible to split indirect
items as well and there were reports on lkml from people who seen this
warning and were worried about it.

To be on the safe side, I consulted with Vladimir Saveliev and he said that
the change is safe to do indeed.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e40996yCDAvDvce5dzZvJmDX2wHg