[PATCH] ppc64: Fix rtas_set_indicator(9005)
It turns out we were passing in the wrong thing to the rtas_set_indicator
call.  Luckily we got away with it because it looks like firmware does not
check arguments and just inserts or removes the current cpu from the global
server group.

Fix it.

Signed-off-by: Anton Blanchard <anton@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e48104d0F-gckjmk-UmVdj2otr7A