[PATCH] kprobes: dont steal interrupts from vm86
The address used by the kprobes handler was not correct if the application
was using LDT entries for code segment.  This patch fixes the above problem
by calculating the address using base address of the current code segment.
Also this patch removes the inline prefix of kprobe_handler() .

Signed-off-by: Prasanna S Panchamukhi <prasanna@in.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e48146ZlyDYilvNlV8sgnFMwGGhQ