[PATCH] pcmcia: remove internal module use count, use module_refcount instead
Remove the internal driver use_count in ds.c, as the reference counting is
done in the module core anyways, and that reference count is available for
cardmgr's usage by a call to module_refcount.  And if !CONFIG_MODULE_UNLOAD,
rmmod is useless anyways, so avoid that call by cardmgr at all.

Signed-off-by: Dominik Brodowski <linux@brodo.de>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e49781SaJ8utFjDXyr2e2yFaMfRA