[PATCH] pcmcia: Add disable_clkrun option
I received report that transmission of Realtek 8139 doesn't work.  The
cause of this problem was CLKRUN protocols of laptop's TI 12xx CardBus
bridge.

CLKRUN is used for controlling the PCI clock by PCI devices (for power saving).
 If device has bug, probably it becomes the cause of stopping the PCI clock at
an unexpected timing.

In the problem of realtek 8139 case, the TX packet which the driver set to the
device was not sent out (still pending state in device).

If "disable_clkrun" is enabled, the problem of TX stopped occurring.  So, I
thought the cause is CLKRUN, and is hardware bug of Cardbus or realtek 8139.
(8139 seems only can control by rewriting EEPROM)

And looks like also
 http://support.microsoft.com/default.aspx?scid=kb;en-us;q294465
 http://support.microsoft.com/default.aspx?scid=kb;en-us;182591
has the problem of CLKRUN.

And I remember that this problem had happened on Thinkpad before.  In the
case, problem seems solved by similar workaround of sound/oss/cs46xx.c.

This patch adds "disable_clkrun" option as workaround of problem to
yenta_socket.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 41e4989eBgbqaHPnTQu3LPhSLGTVVQ