Merge bk://bk.arm.linux.org.uk/linux-2.6-rmk into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/01/14 19:54:18+00:00 catalin.marinas
[ARM PATCH] 2404/1: BTAC/BTB flushing added in cpu_v6_switch_mm

Patch from Catalin Marinas

This is needed since ARMv6 branch prediction tagging is done by the
virtual address and the ASIDs aren't taken into account.

Signed-off-by: Catalin Marinas
Signed-off-by: Russell King

2005/01/14 19:42:46+00:00 ben-linux
[ARM PATCH] 2403/1: S3C2410 - clock initialsation tidy

Patch from Ben Dooks

Update to the s3c2410/s3c2440 clock code to use
the naming conventions now used throughout the
arch/arm/mach-s3c2410/ directory, as well as
a re-organisation of the initialisation of the
clocks.
The previous method was to initialise the clocks
as part of the function that initialised the
statically mapped IO regions, which was called
from the machine initialisation. This proves
a problem if the machine initialisation needs
to use these IO areas to determine the frequency
of PLL crystal fitted.
This patch makes the following changes:
 - s3c2410_init_clocks renamed s3c24xx_setup_clocks
 - s3c2410_clkcon_enable renamed s3c24xx_clk_enable
 - s3c2410_register_clock renamed s3c24xx_register_clock
 - cpu specific init_clocks to be called from cpu.c
 - add s3c24xx_init_clocks for machine init to call
 - split the cpu specific clock code from xxx_map_io
 - update all machine initialisation methods
and the following changes as cleanup which should
have been caught by previous patches:
 - remove s3c2440_hdiv (unused)
 - remove s3c2440_clock_tick_rate (unused)
This patch has been generated from a request by
Dimitry Andric.

Signed-off-by: Ben Dooks
Signed-off-by: Russell King

2005/01/14 19:32:28+00:00 catalin.marinas
[ARM PATCH] 2401/1: asm/thread_info.h removed from arch/arm/vfp/entry.S

Patch from Catalin Marinas

This file no longer needs to be included in arch/arm/vfp/entry.S
because of the inclusion of asm/constants.h.

Signed-off-by: Catalin Marinas
Signed-off-by: Russell King

2005/01/13 22:49:21+00:00 catalin.marinas
[ARM PATCH] 2399/1: asm/constants.h included in arch/arm/vfp/entry.S

Patch from Catalin Marinas

The TI_VFPSTATE is otherwise undefined and gas-2.15.90 (and probably
earlier versions) assumes 0. Gas-2.15.92 reports an error.

Signed-off-by: Catalin Marinas
Signed-off-by: Russell King

2005/01/13 22:43:31+00:00 LW
[ARM PATCH] 2395/1: __ioremap() miscalculates mapping size under certain conditions

Patch from Lothar Wassmann

Hi,
ioremapping a memory range that ends on the first byte of a page
(e.g. mapping PAGE_SIZE + 1 bytes on a page boundary) results in a
mapping that is one page too short. In particular trying to create a 1
byte mapping at the start of a page results in a BUG_ON() assertion in
remap_area_pages() because the calculated mapping size is zero.
This bug has been reported on Bugzilla on 2004-02-25:
http://bugzilla.kernel.org/show_bug.cgi?id=2188

Signed-off-by: Lothar Wassmann
Signed-off-by: Russell King

2005/01/13 22:37:36+00:00 elf
[ARM PATCH] 2394/1: Re: accepting responsibility for Sharp LH ports

Patch from Marc Singer

Admitting that I'm the maintainer for the Sharp LH ports.

Signed-off-by: Marc Singer
Signed-off-by: Russell King

2005/01/13 22:31:41+00:00 rpurdie
[ARM PATCH] 2392/1: Add PCMCIA/CF support code for Sharp SL-C7xx Series

Patch from Richard Purdie

Add support code to enable the Compact Flash socket on the Sharp
SL-C7xx series of PDAs. The socket is controlled via the Sharp
SCOOP interface.

Signed-off-by: Richard Purdie
Signed-off-by: Russell King

BKrev: 41e94ff4GLs5wqyWRtHj9CWizIKfHA